Upgrade swc_ecma_codegen_macros to 0.7.3
I've been chasing some weird dependency schenanigans in one of my projects that causes a ton of syn 2.0 version mismatch errors. I think it comes down to swc doing the syn 2.0 upgrade in a patch release, but given the complexities of the crate graph here I'm not sure.

However, upgrading deno_ast's dependency to 0.7.3 seems to fix it for me.