fix: treat jsx expr tag names as components
This fixes an issue when an expression is passed as the component type like:

```tsx
<Foo.Bar fooBar={2} />
``

Peviously, we wouldn't always treat this as a component and thus do props casing normalization. We must never touch the casing of component props though. Since we can't reliably determine if an expression holds a component or an element, we'll always treat it as a component.